Serve box-score art at clean /mlb/art/<gamePk>.png (proxy, hide bucket)
Was redirecting to the raw Supabase Storage URL (exposed in the address bar
+ shareable links). Now a single same-origin route proxies the bytes so the
URL stays /mlb/art/<gamePk>.png. Blob storage is still the durable cache;
the route reads from it (or generates on first view) and streams the bytes,
immutably CDN-cached. Shareable URL needs no query — the date is only used
to render a 2026 game's header on a first-time miss.
